requirements.txt
Flask
selenium
beautifulsoup4
SQLAlchemy
PyQt5
requests
google-generativeai